requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
psycopg2-binary
Flask-WTF
WTForms
Flask-CKEditor
markdown
bootstrap-flask